requirements.txt
transformers
sentence-transformers
scikit-learn
pandas
numpy
faiss-cpu
streamlit
python-dotenv
